fix repeat operation

+++ canary_reconciler.go
@@ -9,6 +9,7 @@ import (
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -28,7 +29,7 @@ func (r *canaryReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 	var cd canaryv1.Canary
 	err := r.Get(ctx, req.NamespacedName, &cd)
 	if errors.IsNotFound(err) {
-		glog.Error("Could not find canary")
+		glog.Errorf("delete canary: %s, skip", req.NamespacedName)
 		return ctrl.Result{}, nil
 	}
 
@@ -68,6 +69,10 @@ func (r *canaryReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 
 	switch cd.Spec.Stage {
 	case K8sDeployStageCanary:
+		if cd.Status.DeployStatus == K8sDeployStageCanary {
+			glog.Infof("skip canary stage, for: %s", req.Name)
+			return ctrl.Result{}, nil
+		}
 		err = r.Create(ctx, canary)
 		if err != nil {
 			glog.Errorf("unable to create canary deployment of %s: %v", req.Name, err)
@@ -76,20 +81,78 @@ func (r *canaryReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 
 		glog.Infof("create canary deployment name: %s", canaryDeployName)
 
+		// update status
+		cd.Status.DeployStatus = K8sDeployStageCanary
+		if err := r.Update(ctx, &cd); err != nil {
+			glog.Errorf("update canary status err: %s", req.Name)
+		}
+
 	case K8sDeployStageRollBack:
-		r.Delete(ctx, canary)
-		if err != nil {
-			glog.Errorf("unable to delete canary deployment of %s: %v", canaryDeployName, err)
+		if cd.Status.DeployStatus == K8sDeployStageRollBack {
+			glog.Infof("skip rollback stage, for: %s", req.Name)
+			return ctrl.Result{}, nil
+		}
+		var deployment appsv1.Deployment
+		canaryFound := true
+
+		if err := r.Get(ctx, types.NamespacedName{
+			Name: canaryDeployName,
+			Namespace: req.Namespace,
+		}, &deployment); err != nil {
+			if !errors.IsNotFound(err) {
+				glog.Errorf("err can not find canary deployment: %s", canaryDeployName)
+				return ctrl.Result{}, err
+			}
+			canaryFound = false
+		}
+
+		if canaryFound {
+			err = r.Delete(ctx, canary)
+			if err != nil {
+				glog.Errorf("unable to delete canary deployment of %s: %v", canaryDeployName, err)
+				return ctrl.Result{}, err
+			}
+
+			glog.Infof("delete canary deployment name: %s", canaryDeployName)
+		}
+
+
+
+		// update status
+		cd.Status.DeployStatus = K8sDeployStageRollBack
+		if err := r.Update(ctx, &cd); err != nil {
+			glog.Errorf("update canary status err: %s", req.Name)
 			return ctrl.Result{}, err
 		}
 
-		glog.Infof("delete canary deployment name: %s", canaryDeployName)
-
 	case K8sDeployStageRollup:
-		r.Delete(ctx, canary)
-		if err != nil {
-			glog.Errorf("unable to delete canary deployment of %s: %v", canaryDeployName, err)
-			return ctrl.Result{}, err
+		if cd.Status.DeployStatus == K8sDeployStageRollup {
+			glog.Infof("skip rollup stage, for: %s", req.Name)
+			return ctrl.Result{}, nil
+		}
+
+		var deployment appsv1.Deployment
+		canaryFound := true
+
+		if err := r.Get(ctx, types.NamespacedName{
+			Name: canaryDeployName,
+			Namespace: req.Namespace,
+		}, &deployment); err != nil {
+			if !errors.IsNotFound(err) {
+				glog.Errorf("err can not find canary deployment: %s", canaryDeployName)
+				return ctrl.Result{}, err
+			}
+			canaryFound = false
+		}
+
+		if canaryFound {
+			err = r.Delete(ctx, canary)
+			if err != nil {
+				glog.Errorf("unable to delete canary deployment of %s: %v", canaryDeployName, err)
+				return ctrl.Result{}, err
+			}
+
+			glog.Infof("delete canary deployment name: %s", canaryDeployName)
 		}
 
 		for containerName, image := range cd.Spec.Images {
@@ -107,6 +170,12 @@ func (r *canaryReconciler) Reconcile(req ctrl.Request) (ctrl.Result, error) {
 		}
 
 		glog.Infof("update container images : %v of  deployment name: %s", cd.Spec.Images, req.Name)
+
+		// update status
+		cd.Status.DeployStatus = K8sDeployStageRollup
+		if err := r.Update(ctx, &cd); err != nil {
+			glog.Errorf("update canary status err: %s", req.Name)
+		}
 
 	default:
 		glog.Errorf("cannot handle stage %v", cd.Spec.Stage)
